+++ lib/events.js
@@ -14,47 +14,7 @@
 
 const { AdobeIOEvents } = require('@adobe/asset-compute-events-client');
 const fs = require('fs');
-const fetch = require('@adobe/node-fetch-retry');
-
-class AssetComputeWebhookEvents {
-
-    constructor(options) {
-        this.webhookUrl = options.webhookUrl;
-        this.orgId = options.orgId;
-        this.clientId = options.clientId;
-    }
-
-
-
-    async sendEvent(event, retryOptions) {
-        const url = this.webhookUrl;
-        const response = await fetch(url, {
-            method: 'POST',
-            headers: {
-                'Content-Type': 'application/json',
-                'x-ims-org-id': this.orgId,
-                'x-api-key': this.clientId
-            },
-            body: JSON.stringify({
-                user_guid: this.orgId,
-                event_code: event.code,
-                event: Buffer.from(JSON.stringify(event.payload || {})).toString('base64')
-            }),
-            retryOptions: retryOptions
-        });
-        if (response.status !== 200) {
-            // If retry is disabled then anything other than 200 is considered an error
-    
-            // 204 No Content status should be considered a failure by us.
-            // It means that the event was accepted, but there were no registrations interested in
-            // the event. In this case it means that the journal has not been fully created yet.
-            // The result is that the posted event is lost.
-            console.log(`sending event failed with ${response.status} ${response.statusText}`);
-            throw Error(`${response.status} ${response.statusText}`);
-        }
-    }
-    
-}
+const { AssetComputeWebhookEvents } = require('./webhook');
 
 class AssetComputeEvents {
 
@@ -80,7 +40,7 @@ class AssetComputeEvents {
             fs.mkdirSync(this.fsEventPath, { recursive: true });
             this.fsEventCounter = 0;
         } else if(params.auth && params.auth.isServiceAccount && params.auth.webhookUrl) {
-            console.log(`$$$ initiating AssetComputeWebhookEvents to send events to ${params.auth.webhookUrl}`);
+            console.log(`Initiating AssetComputeWebhookEvents to send events to ${params.auth.webhookUrl}`);
             this.webhookEvents = new AssetComputeWebhookEvents({
                 webhookUrl: params.auth.webhookUrl,
                 orgId: orgId,
@@ -88,7 +48,7 @@ class AssetComputeEvents {
             });
 
         } else if (params.auth && params.auth.accessToken && orgId && clientId) {
-            console.log(`$$$ initiating AdobeIOEvents to send events `);
+            console.log('Initiating AdobeIOEvents to send events');
             this.ioEvents = new AdobeIOEvents({
                 accessToken: params.auth.accessToken,
                 orgId: orgId,
@@ -127,31 +87,26 @@ class AssetComputeEvents {
                 fs.writeFileSync(eventJsonPath, JSON.stringify(payload));
                 console.log(`successfully wrote event ${type} as ${eventJsonPath}`);
 
+            } else if(this.webhookEvents) {
+                console.log(`Using webhook to send event ${type} to ${this.webhookEvents.webhookUrl}`);
+                await this.webhookEvents.sendEvent({
+                    code: AssetComputeEvents.EVENT_CODE,
+                    payload: payload
+                }, this.retry);
+                console.log("Webhook : successfully sent event");
+            } else if (this.ioEvents) {
+                console.log(`Using IO Events sending event ${type} as ${this.getProviderId()}`);
+                await this.ioEvents.sendEvent({
+                    code: AssetComputeEvents.EVENT_CODE,
+                    payload: payload
+                }, this.retry);
+                console.log("IO Events : successfully sent event");
             } else {
-                if (!this.ioEvents && !this.webhookEvents) {
-                    // Logging info about event is useful when running in test environments
-                    console.log("event not sent:", type, JSON.stringify(payload));
-                    return;
-                }
-                if(this.webhookEvents) {
-                    console.log(`$$$$ Using webhook to send event ${type} as ${this.getProviderId()}`);
-                    await this.webhookEvents.sendEvent({
-                        code: AssetComputeEvents.EVENT_CODE,
-                        payload: payload
-                    }, this.retry);
-                    console.log("$$$$ Webhook : successfully sent event");
-                } else {
-
-                    console.log(`sending event ${type} as ${this.getProviderId()}`);
-                    console.log(`$$$$ Using IO Events to send event ${type} as ${this.getProviderId()}`);
-                    await this.ioEvents.sendEvent({
-                        code: AssetComputeEvents.EVENT_CODE,
-                        payload: payload
-                    }, this.retry);
-                    console.log("$$$$ IO Events : successfully sent event");
-                }
-
+                // Logging info about event is useful when running in test environments
+                console.log("event not sent:", type, JSON.stringify(payload));
+                return;
             }
+
         } catch (e) {
             if(params && params.requestId){
                 console.error(`error sending event for requestId ${params.requestId}: ${e.message}, (details: ${JSON.stringify(e)})`);
